pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Vyacheslav <danubise@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
uvicorn = "^0.27.0.post1"
gunicorn = "^21.2.0"
pyyaml = "^6.0.1"
fastapi = "^0.109.0"
pandas = "^2.2.0"
transformers = "^4.37.2"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
